+++ Controllers/StockController.cs
@@ -21,14 +21,22 @@ namespace InformationSystems.Server.Controllers
 
         [HttpGet]
         public async Task<IActionResult> GetAll() {         // mapper
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var stocks = await _stockRepo.GetAllAsync();
             var stocksDTO = stocks.Select(x => x.ToStockDTO());
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
-        {   
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var stock = await _stockRepo.GetByIdAsync(id);
             if (stock==null)
             {
@@ -40,15 +48,23 @@ namespace InformationSystems.Server.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequest stockDTO)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var stock = stockDTO.ToStockFromCreateDTO();
             await _stockRepo.CreateAsync(stock);
             return CreatedAtAction(nameof(GetById), new { id = stock.Id }, stock.ToStockDTO());
         }
 
         [HttpPut]
-        [Route("/api/stock/{id}")]
+        [Route("/api/stock/{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequest stockDTO)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var stock = await _stockRepo.UpdateAsync(id, stockDTO);
             if (stock == null)
             {
@@ -58,9 +74,13 @@ namespace InformationSystems.Server.Controllers
         }
 
         [HttpDelete]
-        [Route("/api/stock/{id}")]
+        [Route("/api/stock/{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var stock = await _stockRepo.DeleteAsync(id);
             if (stock == null)
             {
